+++ offer-manager-domain/src/main/groovy/life/qbic/business/offers/OfferItem.groovy
@@ -17,6 +17,8 @@ class OfferItem {
     final String productName
     final double unitPrice
     final double quantityDiscount
+    final double discountPerUnit
+    final double discountPercentage
     final String serviceProvider
     final String unit
     final double itemTotal
@@ -27,19 +29,23 @@ class OfferItem {
         String productName
         double unitPrice
         double quantityDiscount
+        double discountPerUnit
+        double discountPercentage
         String serviceProvider
         String unit
         double itemTotal
 
-        Builder(double quantity, String productDescription, String productName, double unitPrice, double quantityDiscount, String serviceProvider, String unit, double itemTotal) {
-            this.quantity =  Objects.requireNonNull(quantity, "Customer must not be null")
-            this.productDescription =  Objects.requireNonNull(productDescription, "Customer must not be null")
-            this.productName =  Objects.requireNonNull(productName, "Customer must not be null")
-            this.unitPrice =  Objects.requireNonNull(unitPrice, "Customer must not be null")
-            this.quantityDiscount =  Objects.requireNonNull(quantityDiscount, "Customer must not be null")
-            this.serviceProvider =  Objects.requireNonNull(serviceProvider, "Customer must not be null")
-            this.unit =  Objects.requireNonNull(unit, "Customer must not be null")
-            this.itemTotal =  Objects.requireNonNull(itemTotal, "Customer must not be null")
+        Builder(double quantity, String productDescription, String productName, double unitPrice, double quantityDiscount, double discountPerUnit, double discountPercentage, String serviceProvider, String unit, double itemTotal) {
+            this.quantity =  Objects.requireNonNull(quantity, "Quantity must not be null")
+            this.productDescription =  Objects.requireNonNull(productDescription, "Product description must not be null")
+            this.productName =  Objects.requireNonNull(productName, "Product name must not be null")
+            this.unitPrice =  Objects.requireNonNull(unitPrice, "Unit price must not be null")
+            this.quantityDiscount =  Objects.requireNonNull(quantityDiscount, "Quantity discount must not be null")
+            this.discountPerUnit = Objects.requireNonNull(discountPerUnit, "Discount per unit must not be null")
+            this.discountPercentage = Objects.requireNonNull(discountPercentage, "Discount percentage must not be null")
+            this.serviceProvider =  Objects.requireNonNull(serviceProvider, "Service provider must not be null")
+            this.unit =  Objects.requireNonNull(unit, "Unit must not be null")
+            this.itemTotal =  Objects.requireNonNull(itemTotal, "Item total must not be null")
 
         }
 
@@ -54,6 +60,8 @@ class OfferItem {
         this.productName = builder.productName
         this.unitPrice = builder.unitPrice
         this.quantityDiscount = builder.quantityDiscount
+        this.discountPerUnit = builder.discountPerUnit
+        this.discountPercentage = builder.discountPercentage
         this.serviceProvider = builder.serviceProvider
         this.unit = builder.unit
         this.itemTotal = builder.itemTotal