+++ offer-manager-domain/src/test/groovy/life/qbic/portal/portlet/offers/OfferSpec.groovy
@@ -1,5 +1,6 @@
 package life.qbic.portal.portlet.offers
 
+import life.qbic.business.ProductFactory
 import life.qbic.business.offers.Offer
 import life.qbic.business.offers.QuantityDiscount
 import life.qbic.business.offers.identifier.OfferId
@@ -146,8 +147,8 @@ class OfferSpec extends Specification {
         double totalDiscount = new QuantityDiscount().apply(primaryAnalysisItem.quantity as Integer,
                 (primaryAnalysisItem.product.externalUnitPrice * primaryAnalysisItem.quantity) as BigDecimal)
         double expectedNetSum = (10.0 + (400 * 1.0))
-        double expectedOverhead = (10.0 + (400 * 1.0) - totalDiscount) * 0.2
-        double expectedTaxes = (expectedNetSum + expectedOverhead) * 0.19
+        double expectedOverhead = (expectedNetSum - totalDiscount) * 0.2
+        double expectedTaxes = (expectedNetSum + expectedOverhead - totalDiscount) * 0.19
         offer.items.size() == 2
         netSum == expectedNetSum
         overhead == expectedOverhead
@@ -181,7 +182,7 @@ class OfferSpec extends Specification {
         then:
         double expectedNetSum = (10.0 + (400 * 1.0))
         double expectedOverhead = (10.0 + (400 * 1.0) - totalDiscount) * 0.4
-        double expectedTaxes = (expectedNetSum + expectedOverhead) * 0.19
+        double expectedTaxes = (expectedNetSum + expectedOverhead - totalDiscount) * 0.19
         offer.items.size() == 2
         netSum == expectedNetSum
         overhead == expectedOverhead
@@ -456,13 +457,68 @@ class OfferSpec extends Specification {
         List<ProductItem> items = [primaryAnalysis, projectManagement, sequencing, dataStorage, secondaryAnalysis]
         Offer offer = new Offer.Builder(customerWithAllAffiliations, projectManager, "Awesome Project", "An " +
                 "awesome project", items, affiliation).build()
-        double netPrice
+        double expectedResult = items.sum {(it.quantity * it.product.externalUnitPrice) as double}
 
-        when: "the net price is calculated"
-        netPrice = offer.getTotalNetPrice()
+        expect: "the calculated costs equal the expected costs"
+        offer.getTotalNetPrice() == expectedResult
 
-        then: "the correct prices are taken into account"
-        netPrice == items.sum {(it.quantity * it.product.externalUnitPrice) as double}
+        where: "the affiliation is"
+        affiliation << [externalAffiliation, externalAcademicAffiliation]
+    }
+
+
+    /**
+     * @since 1.1.0
+     */
+    def "the total net costs with overheads are computed with the correct internal prices"() {
+        given: "a list of product items with internal and external base prices"
+        ProductItem primaryAnalysis = new ProductItem(2, new PrimaryAnalysis("Basic RNAsq", "Just an" +
+                " example primary analysis", 2.5, 3.5, ProductUnit.PER_SAMPLE, 1, Facility.QBIC))
+        ProductItem secondaryAnalysis = new ProductItem(1, new SecondaryAnalysis("Basic RNAsq", "Just an" +
+                " example secondary analysis", 2.4, 42.56, ProductUnit.PER_SAMPLE, 1, Facility.QBIC))
+        ProductItem sequencing = new ProductItem(3, new Sequencing("Basic Sequencing", "Just an" +
+                "example sequencing", 3.0, 4.6, ProductUnit.PER_SAMPLE, 1, Facility.QBIC))
+        ProductItem projectManagement = new ProductItem(1, new ProjectManagement("Basic Management",
+                "Just an example", 10.0, 11.26, ProductUnit.PER_DATASET, 1, Facility.QBIC))
+        ProductItem dataStorage = new ProductItem(2, new DataStorage("Data Storage",
+                "Just an example", 20.0, 23, ProductUnit.PER_DATASET, 1, Facility.QBIC))
+        and: "an offer with these items"
+        List<ProductItem> items = [primaryAnalysis, projectManagement, sequencing, dataStorage, secondaryAnalysis]
+        Offer offer = new Offer.Builder(customerWithAllAffiliations, projectManager, "Awesome Project", "An " +
+                "awesome project", items, affiliation).build()
+        double expectedResult = items.sum {(it.quantity * it.product.internalUnitPrice) as BigDecimal} - offer.getTotalDiscountAmount() + offer.getOverheadSum()
+
+        expect: "the calculated costs equal the expected costs"
+        offer.getTotalNetPriceWithOverheads() == expectedResult
+
+        where: "the affiliation is"
+        affiliation << [internalAffiliation]
+
+    }
+
+    /**
+     * @since 1.1.0
+     */
+    def "the total net costs with overheads are computed with the correct external prices"() {
+        given: "a list of product items with internal and external base prices"
+        ProductItem primaryAnalysis = new ProductItem(2, new PrimaryAnalysis("Basic RNAsq", "Just an" +
+                " example primary analysis", 2.5, 3.5, ProductUnit.PER_SAMPLE, 1, Facility.QBIC))
+        ProductItem secondaryAnalysis = new ProductItem(1, new SecondaryAnalysis("Basic RNAsq", "Just an" +
+                " example secondary analysis", 2.4, 42.56, ProductUnit.PER_SAMPLE, 1, Facility.QBIC))
+        ProductItem sequencing = new ProductItem(3, new Sequencing("Basic Sequencing", "Just an" +
+                "example sequencing", 3.0, 4.6, ProductUnit.PER_SAMPLE, 1, Facility.QBIC))
+        ProductItem projectManagement = new ProductItem(1, new ProjectManagement("Basic Management",
+                "Just an example", 10.0, 11.26, ProductUnit.PER_DATASET, 1, Facility.QBIC))
+        ProductItem dataStorage = new ProductItem(2, new DataStorage("Data Storage",
+                "Just an example", 20.0, 23, ProductUnit.PER_DATASET, 1, Facility.QBIC))
+        and: "an offer with these items"
+        List<ProductItem> items = [primaryAnalysis, projectManagement, sequencing, dataStorage, secondaryAnalysis]
+        Offer offer = new Offer.Builder(customerWithAllAffiliations, projectManager, "Awesome Project", "An " +
+                "awesome project", items, affiliation).build()
+        double expectedResult = items.sum {(it.quantity * it.product.externalUnitPrice) as double} - offer.getTotalDiscountAmount() + offer.getOverheadSum()
+
+        expect: "the calculated costs equal the expected costs"
+        offer.getTotalNetPriceWithOverheads() == expectedResult
 
         where: "the affiliation is"
         affiliation << [externalAffiliation, externalAcademicAffiliation]
@@ -605,77 +661,5 @@ class OfferSpec extends Specification {
         10 | 0.67
         100 | 0.3
 
-    }
-
-    static class ProductFactory {
-        static <T extends Product> T createProduct(Class clazz, String description, String name, double internalPrice, double externalPrice, Facility serviceProvider) {
-            int runningNumber = 1
-            ProductUnit productUnit = ProductUnit.PER_SAMPLE
-            switch (clazz) {
-                case DataStorage:
-                    return new DataStorage(name,
-                            description,
-                            internalPrice,
-                            externalPrice,
-                            productUnit,
-                            runningNumber,
-                            serviceProvider) as T
-                    break
-                case MetabolomicAnalysis:
-                    return new MetabolomicAnalysis(name,
-                            description,
-                            internalPrice,
-                            externalPrice,
-                            productUnit,
-                            runningNumber,
-                            serviceProvider) as T
-                    break
-                case PrimaryAnalysis:
-                    return new PrimaryAnalysis(name,
-                            description,
-                            internalPrice,
-                            externalPrice,
-                            productUnit,
-                            runningNumber,
-                            serviceProvider) as T
-                    break
-                case ProjectManagement:
-                    return new ProjectManagement(name,
-                            description,
-                            internalPrice,
-                            externalPrice,
-                            productUnit,
-                            runningNumber,
-                            serviceProvider) as T
-                    break
-                case ProteomicAnalysis:
-                    return new ProteomicAnalysis(name,
-                            description,
-                            internalPrice,
-                            externalPrice,
-                            productUnit,
-                            runningNumber,
-                            serviceProvider) as T
-                    break
-                case SecondaryAnalysis:
-                    return new SecondaryAnalysis(name,
-                            description,
-                            internalPrice,
-                            externalPrice,
-                            productUnit,
-                            runningNumber,
-                            serviceProvider) as T
-                    break
-                case Sequencing:
-                    return new Sequencing(name,
-                            description,
-                            internalPrice,
-                            externalPrice,
-                            productUnit,
-                            runningNumber,
-                            serviceProvider) as T
-                    break
-            }
-        }
     }
 }